requirements.txt
langchain==0.2.0
langchain_openai==0.1.7
python-dotenv==1.0.1
fastapi==0.110.2
uvicorn==0.29.0
redis==5.0.4
langchainhub==0.1.15
langgraph==0.0.50
langchain-core==0.2.0
langchain-community==0.2.0
langchain-elasticsearch==0.2.1
tiktoken==0.7.0
rapidocr-onnxruntime==1.2.13
google-api-python==2.129.0
google-auth==2.29.0
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
unstructured==0.12.6
colorlog==6.8.2